requirements.txt
Django==1.5.2
South==0.8.2
Pillow==2.1.0
celery==3.0.23
kombu==2.5.14
django-celery==3.0.23
django-model-utils==1.5.0
django-userena==1.2.3
django-uuidfield==0.4.0
django-storages==1.1.8
django-configurations==0.4
docutils==0.11
eventlet==0.13.0
gunicorn==18.0
netaddr==0.7.10
requests==1.2.3
djangorestframework==2.3.6